add helpers and tests for orgs and related names
This gives us some convenience methods,
allowing us to call (for example)

`org.get_official_name(date(2017, 12, 1))` or
`org.get_slug(datetime.today())`

Given that in 99% of cases there is a one-to-one
mapping between Organisation and OrganisationName
(this is the most common case by a long way)
I've chosen to just return the one name in that
case, without checking the dates.

That allows us to just assume the Org start and
end dates are the same for the one associated name.

Also, on an ad-hoc basis (e.g: on the console),
if we know we don't need to deal with multiple names,
we can juse use:

`org.get_slug()` as a shorthand for

```
from datetime import datetime
org.get_slug(datetime.today())
```

which is handy.